Added Template editing. Creating with templates not supported yet.
Templates are listed in Context-Menu under the AddEntry Toolbar
Toolbar Button is missing an Arrow for now. Control-Size is not working correctly for now
Nested Template or Trash groups aren't considered, this is a bug!
Minor changes to the UI (Settings tabs now use common icons)
Added Workflow-Settings tab to extract all the custom action possible on entries. The copy or open on URL dbl-click setting will move over to this tab.