MPToolbarDelegate now takes care of the toolbar manipulation.
Toolbar is shown when search is started if it was hidden. If search revealed the toolbar, it will be hidden after the search is exited
If the toolbar was using text only mode, it switches to icon/text mode and then back to text mode after the search - there might be a better way to deal with this by simply using a menu for the search.