New & old icons
@mstarke I tried to make the icons simple as well (not much else to
strive for with 16x16 pixels and the color black, he he).

Yes, some of the symbols just doesn't play well with the overall
concept of getting them to "feel" the same. The first try was just
that. Now I have had time to work some more with the icons and hope it
is for the better.

I'm thinking working with circles rather than rounded squares might
work better regarding making them take up approximately the same space.

Some of the icons I inverted using a black circle like the W and the %,
to make them seem bigger. W might be Wikipedia or maybe Wordpress so I
kept it as a letter rather than go with a puzzle logo, although we
could try with that as well. Is there any limit on how many icons we
can have???

I changed the open folder to be more straight on open. I made two new
icons for number 47 (the package icon ... yes, that one was bad), one
would be a traditional compress symbol and the other a package with a
belt on ... does it show?

Instead of the tilted mail on number 25 I made a mailbox.Number 66 got
converted to three bills.

I didn't come up with anything revolutionary regarding the battery, but
I do think it could work with a view straight on as well.

@DominusExult I got all the native Keepass icons from mstarke and have
absolutely no idea from which version it is from. I have made another
version now ... this time it is a life belt instead, since it's
supposed to stand for help.

I'm not entirely sure how I feel about the file save icon (26). Most
people might know about the floppy disk and the arrow I associate more
with dowmloading than saving, but for now I have kept it as is. I might
make a floppy disk as well at some point.

Icon 0 was not placed first because og the way Illustrator works (there
is no artboard number 0, so I opted to make it number 81 (which was the
last at the time), however, now I have made it number 100 and placed it
first now that I rearranged the whole setup with 10 icons in each row
and numbers to make it easier to refer to an icon (in case you do not
recognise what it is). So I didn't forget the key icon ... I just
placed it last.

It is the first time I have worked on a GitHub project, so I hope I got
things right.