Update on icons
Thank you for all your notes @mstarke ... it's nice to have something
to work with.

* Cool ... I changed all the keys except the one in the back in 13
* 16 So I mirrored the needle, however, I disagree with you on the
point that it would fit with the CD better. First of it has a different
angle which just makes it seem off and changing the angle of either
looks weird (to me at least) and secondly I rather like, that they are
not too similar. We want people to be able to distinguish between the
icon easily ... right?!
* 25 I did think I nailed it pretty good when had done this one ...
nice to know you agree :)
* 26 Moved the save on harddisk to 89 and made a floppy disk.
* 38 I have no clue what Samba is so I just went with the original
icon. Do tell if you wish something else.
* 32 Filesystem viwer ... I think you are right. I alterede the icon
plus I found a place where they describe fsview as follows:

>       FSView is a tool for showing disc utilization in a graphical
form, much
>       like the UNIX  command  ’du’.   The  visualisation  type
chosen  is  a
       treemap.   Each  directory  is represented by a rectangle whose
area is
       proportional to its size.  Each directory  rectangle  contains
smaller
       rectangles for its contents, and so on down.

* 33 I guess someone who love the race track ...
* 42 this was problematic because for some reason the little indents
does not show when the bar is not tilted, so I had to make them bigger,
which makes them look stupid, when the icon is bigger and for some
reason they are more recognisable tilted ... what do you guys think?
* 47 So I made a third option ... a zipped circle.
* 49 It might work with the perspective.
* 58 Is the key too big if it is just put in full size?
* 62 Hi hi ... thanks :D
* 66 One $sign coming up.

Before I add the .ai file I would like us to get a little closer to
what we will end up with, since the .ai file is a big mess right now
... a lot of the icons have not been expanded yet and made into
"proper" vector shapes fit for your purpose. I won't do that until I am
fairly certain we have no more tweaks to make.