Registered to closeGrouping on undo manger to update searches on changes
This fixes issues with search not updating when something on the data model is changes e.g. via drag and drop. Also this fixes a display error that occured when an entry was dragged into the selected group in the outline view which would then show the group contents and not the search results